ME-637: Renamed document.py to "documents.py".
